Move check for push to AppSec service init listener
